+++ lib/ui/emulator/debug_overlay.g.dart
@@ -1,0 +1,45 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'debug_overlay.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$debugOverlayControllerHash() =>
+    r'05fcdcb58929c3fd4af7c85f50a19c0e8b25e73d';
+
+/// See also [debugOverlayController].
+@ProviderFor(debugOverlayController)
+final debugOverlayControllerProvider =
+    AutoDisposeProvider<DebugOverlayController>.internal(
+  debugOverlayController,
+  name: r'debugOverlayControllerProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$debugOverlayControllerHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef DebugOverlayControllerRef
+    = AutoDisposeProviderRef<DebugOverlayController>;
+String _$debugOverlayNotifierHash() =>
+    r'6baa34aac142c2298d7490b6a5b0ba0c529fbb9d';
+
+/// See also [DebugOverlayNotifier].
+@ProviderFor(DebugOverlayNotifier)
+final debugOverlayNotifierProvider = AutoDisposeNotifierProvider<
+    DebugOverlayNotifier, DebugOverlayState>.internal(
+  DebugOverlayNotifier.new,
+  name: r'debugOverlayNotifierProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$debugOverlayNotifierHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$DebugOverlayNotifier = AutoDisposeNotifier<DebugOverlayState>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member